alx_travel_app/alx_travel_app/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
<<<<<<< HEAD:alx_travel_app/alx_travel_app/requirement.txt
django-cors-headers==4.7.0
=======
django-cors-headers==4.9.0
>>>>>>> 1ad8b2d2c8aabdac814ec6b09ff6835e35355991:alx_travel_app/requirements.txt
django-environ==0.12.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.4
<<<<<<< HEAD:alx_travel_app/alx_travel_app/requirement.txt
mysqlclient==2.2.7
=======
>>>>>>> 1ad8b2d2c8aabdac814ec6b09ff6835e35355991:alx_travel_app/requirements.txt
packaging==25.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
